[make] Update interpreter to work at non-identity mapped addresses
Details

  On Arm-based MacOS, the first 4Gb of address space, called the zero
  page (or Page Zero) are off limits to a user program.

  The current implementation of this Oberon interpreter uses mmap() to
  map the Oberon heap below 4Gb.  On Linux, for simplicity, it is
  mapped at the same address to which Oberon is loaded.  This means
  that Oberon addresses and host addresses have the same value.  But,
  because it is mapped below 4Gb, this method cannot be used on
  Arm-based MacOS.  Instead, the memory must be mapped above 4Gb.  The
  main constraint for the mapping is that bit 30 and bit 31 of the
  Oberon address space must be 0.

  loader/fileutils.cpp

    Clang provides better type security than gcc.  The type changes
    here are to fix issues reported by Clang on MacOS.

  loader/heap.cpp
  loader/heap.h

  loader/loader.cpp

    Changes here fix type issues not reported by gcc, and re-adds the
    ability to dump the Oberon heap on exit (don't know when that
    capability got lost).

  loader/o3.cpp

    All accesses to a module name were incorrectly managed because the
    Oberon heap and C++ program address space were the same.  Now that
    the address space is guaranteed to be different, accessing the
    module name requires conversion to a host address from the Oberon
    heap address.  The helper function 'module_name' does that.

    Additionally, new diagnostic messages have been added to track the
    progress of bootstrapping the Oberon heap.

  loader/o3.h

    Change signature of 'module_name' function to not return a const
    pointer; this makes it easier to deal with both Oberon and C++
    types.

  loader/skl.h

   Fix type issues revealed by Clang, and add another invariant assert
   in aligned().

  loader/skl_instruction.cpp

   Fix type issues revealed by Clang.

  loader/skl_instruction.h

   Fix type issues revealed by Clang.

Testing

  skl-test-compiler

   Successful.

  skl-test-modules

   Successful.

  skl-test-system

    Successful.

+++ loader/skl_instruction.h
@@ -1,4 +1,4 @@
-/* Copyright (c) 2022 Logic Magicians Software */
+/* Copyright (c) 2022-2026 Logic Magicians Software */
 #if !defined(_SKL_INSTRUCTION_H)
 #define _SKL_INSTRUCTION_H
 
@@ -35,7 +35,7 @@ namespace skl {
     };
 
     extern instruction_t **cache;
-    extern int             cache_elements;
+    extern unsigned long   cache_elements;
 
     bool allocate_instruction_cache(int heap_mb, int stack_mb);
     void release_instruction_cache(void);
@@ -44,7 +44,7 @@ namespace skl {
     static inline instruction_t *
     lookup_instruction(md::OADDR addr)
     {
-        int offset = heap::heap_offset(addr);
+        unsigned offset = heap::heap_offset(addr);
 
         assert(((offset & (static_cast<int>(sizeof(md::uint32)) - 1)) == 0) &&
                offset < cache_elements);
